refactor(pdf): isolate bounded table scope matching
